feat: 주문 API 구현

+++ src/main/java/com/team/saver/market/coupon/dto/CouponCreateRequest.java
@@ -9,7 +9,7 @@ public class CouponCreateRequest {
 
     private String couponDescription;
 
-    private double saleRate;
+    private int saleRate;
 
     private long marketId;
